tests(unit): move compatibility tests to become unit tests for grid
