+++ src/tests/integration/mod.rs
@@ -1,5 +1,4 @@
 pub mod close_pane;
-pub mod compatibility;
 pub mod layouts;
 pub mod move_focus_down;
 pub mod move_focus_left;
